requirements.txt
aiomysql==0.2.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==46.0.2
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
openpyxl==3.1.5
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
resend==2.17.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
